Se termino empresas y se agrego persona por ciudad y personas en varias ciudades

+++ src/main/java/com/besysoft/agenda/ServiceImp/EmpresaService.java
@@ -3,6 +3,7 @@ package com.besysoft.agenda.ServiceImp;
 import com.besysoft.agenda.Entity.Empresa;
 import com.besysoft.agenda.Exception.BadRequestException;
 import com.besysoft.agenda.Exception.RecursoNoEncontradoException;
+import com.besysoft.agenda.IService.IContactoService;
 import com.besysoft.agenda.IService.IEmpresaService;
 import com.besysoft.agenda.Repository.EmpresaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +18,9 @@ public class EmpresaService implements IEmpresaService {
 
     @Autowired
     private EmpresaRepository empresaRepository;
+
+    @Autowired
+    private IContactoService contactoService;
 
     @Override
     public Empresa obtenerEmpresa(Long id) {
@@ -47,6 +51,9 @@ public class EmpresaService implements IEmpresaService {
     @Override
     public void eliminarEmpresa(Long id) {
         Empresa empresa = obtenerEmpresa(id);
+        if(contactoService.existeContactoPorEmpresa(empresa.getId())){
+            throw new BadRequestException("La empresa " + empresa.getNombre() + "tiene referencias en contactos y no se puede eliminar.");
+        }
         empresaRepository.deleteById(empresa.getId());
     }
 }
